+++ frontend-gd/src/main/modules/nav-bar/logged-in-navbar.js
