+++ frontend-gd/src/main/modules/nav-bar/navbar-bootstrap.js
