+++ frontend-gd/src/main/modules/table-module/table-module.css
