--- frontend-gd/src/main/pages/admin-login/index.js
@@ -1,1 +1,0 @@
-export { default } from './admin-login';
