+++ frontend-gd/src/main/pages/date-generator-page/date-generator.js
