+++ frontend-gd/src/main/pages/date-generator-page/index.js
