+++ frontend-gd/src/main/pages/log-in-page/log-in-page.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react';
 import variables from '../../tools/variables'
 
 import LogIn from '../../components/log-in';
-import NavbarBootstrap from '../../components/navbar-bootstrap';
+import NavbarBootstrap from '../../modules/NavBar/navbar-bootstrap';
 
 import '../../../App.scss';
