+++ frontend-gd/src/main/tools/registration-store.js
